requirements.txt
gradio==3.50.2
torch==2.1.2+cpu
torchvision==0.16.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
diffusers==0.27.2
transformers==4.38.2
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
accelerate==0.27.2
safetensors==0.4.2
huggingface_hub==0.20.3
pydantic==1.10.13 